requirements.txt
autopep8==2.0.2
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
tomli==2.0.1
Werkzeug==2.3.3
